Consolidate team-invites-schema into team-schema
