+++ src/routes/@shared/index.ts
@@ -1,20 +1,17 @@
 export { rules } from '../rules'
 
 export {
-  type InviteMemberBody,
-  inviteMemberBodySchema,
-  type InviteMemberCtx,
-  type TeamParams as TeamInvitesParams,
-  teamParamsSchema as teamInvitesParamsSchema,
-} from './team-invites-schema'
-
-export {
   type CreateTeamBody,
   createTeamBodySchema,
   type CreateTeamCtx,
   type GetTeamsCtx,
   type GetTeamsQuery,
   getTeamsQuerySchema,
+  type InviteMemberBody,
+  inviteMemberBodySchema,
+  type InviteMemberCtx,
+  type TeamParams as TeamInvitesParams,
+  teamParamsSchema as teamInvitesParamsSchema,
   type TeamParams,
   type TeamParamsCtx,
   teamParamsSchema,
