+++ src/routes/rules.ts
@@ -65,6 +65,7 @@ export const rules = {
   },
 
   team: {
+    id: z.uuid(),
     name: z.string().trim().min(1).max(255),
     website: z.url().max(255),
     description: z.string().trim().max(2000),
